% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CampbellAllImport.R
\name{CampbellAllImport}
\alias{CampbellAllImport}
\title{Import several Campbell Science AgFace data logger *.dat files from a Folder}
\usage{

  CampbellAllImport(logger.folder = "~/AgFace/2015/Campbell_logger/logger_data",
  log.interval = "5Min", logger.name = NA, checkduplicates = TRUE)
}
\arguments{
\item{logger.folder}{Full path to the location of the *.dat files. Defaults to "~/AgFace/2015/Campbell_logger/logger_data"}

\item{log.interval}{Select the files based on the common log interval that is specified in the filename. Allowed values are \code{"5Min", "15Min", "Hourly", "Daily"}. Only one time interval is allowed.}

\item{logger.name}{Select files from one logger only. \code{logger.name} has to match the System name as given in the file name. Allowed values are \code{"SYS1", "SYS2", ..., "SYS8"}}
}
\value{
Returns data frame with all imported *.dat files.
}
\description{
Uses \code{\link{CampbellFileImport}} to import all *.dat files from a folder. Files are selected based on their log-interval that is part of their file name. Allowed intervals are "5Min", "15Min", "Hourly", "Daily".
}

