% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampbellLogger-package.R
\docType{package}
\name{CampbellLogger}
\alias{-package}
\alias{CampbellLogger}
\title{Import and process Campbell logger data}
\description{
Import single or multiple Campbell logger *.dat files.
}
