% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetSensorID.R
\name{GetSensorID}
\alias{GetSensorID}
\title{Identify the SensorID associated with a specific sensor}
\usage{
GetSensorID(sensor.name)
}
\arguments{
\item{sensor.name}{Fullname of the Sensor from the header of the dat file. Has to be part of the list \code{sensor.names}}
}
\value{
Data frame with Fullname, SensorId, and Sensor name
}
\description{
Identify the SensorID associated with a specific sensor
}

