% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MyKshPlot.R
\name{MyKshPlot}
\alias{MyKshPlot}
\title{Visualise the Sap flow paramter Ksh for a given date}
\usage{
MyKshPlot(data, date = Sys.time(), para = "Kshapp_Avg", ylim = c(0, 1))
}
\arguments{
\item{data}{Data frame name with the data}

\item{date}{Character string in POSIXct format (YYYY-mm-dd). Defaults to the current date.}

\item{para}{The parameter to visualise. Defaults to "Kshapp_Avg".}

\item{ylim}{y-axis range as vector. Defautls to \code{c(0, 1)}}
}
\value{
Returns a ggplot object.
}
\description{
Version of MyRecentPlot with changes specific to visualising the thermal conductance constant \code{KSh} for a given spa flow gauge installation. Visualises Ksh over the time from 4:00 to 6:00 hrs for the given date with a y-axis range from 0 to 1.
}

