% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CampbellLocations.R
\docType{data}
\name{Agface.loc}
\alias{Agface.loc}
\title{Locations of AgFace and Creswick campus site}
\format{\preformatted{ num [1, 1:2] 142.1 -36.8
}}
\usage{
Agface.loc
}
\description{
Code to provide locations for Agface and Creswick in format needed for \code{CampbellSunriseSunset}. Agface field site position -36.751367, 142.114477, Creswick position -37.423003, 143.900472
}
\keyword{datasets}

