% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CampbellSunriseSunset.R
\name{CampbellSunriseSunset}
\alias{CampbellSunriseSunset}
\title{Calculate sunrise and sunset times based on location and date for Campell *.dat files}
\usage{
CampbellSunriseSunset(data, location = Agface.loc, DayLightSaving = FALSE)
}
\arguments{
\item{data}{Data frame with a TIMESTAMP entry}

\item{location}{Location of the measurement for which to calculate sunrise and sunset. Given in matrix format containing x and y coordinates. Defaults to \code{Agface.loc.}}

\item{DayLigthSaving}{Logical. If \code{TRUE} one hour is subtracted from the calculated sunrise and sunset times.}
}
\description{
Calculate sunrise and sunset times based on location and date for Campell *.dat files
}

